@@ -85,7 +85,7 @@ class ArrayUserRepository implements UserRepository
         $refl = new \ReflectionClass($user);
         $id   = $refl->getProperty('id');
         $id->setAccessible(TRUE);
-        $id->setValue($user, uniqid());
+        $id->setValue($user, \uniqid());
     }
 
     /**
