@@ -11,7 +11,7 @@ class FixedUrlProviderStub implements UrlProvider
 {
     public function getAfterLoginUrl(User $user)
     {
-        return '/after-login?'.http_build_query(['email' => $user->getEmail()]);
+        return '/after-login?'.\http_build_query(['email' => $user->getEmail()]);
     }
     public function getAfterLogoutUrl()
     {
@@ -35,35 +35,35 @@ class FixedUrlProviderStub implements UrlProvider
 
     public function getCompleteActivationUrl(array $params)
     {
-        return '/complete-activation?'.http_build_query($params);
+        return '/complete-activation?'.\http_build_query($params);
     }
 
     public function getCompleteChangeEmailUrl(array $params)
     {
-        return '/complete-change-email?'.http_build_query($params);
+        return '/complete-change-email?'.\http_build_query($params);
     }
 
 
     public function getCompletePasswordResetUrl(array $params)
     {
-        return '/complete-password-reset?'.http_build_query($params);
+        return '/complete-password-reset?'.\http_build_query($params);
     }
 
     public function getCompleteRegistrationUrl(array $params)
     {
-        return '/complete-registration?'.http_build_query($params);
+        return '/complete-registration?'.\http_build_query($params);
     }
 
     public function getDefaultUserHomeUrl(User $user)
     {
-        return '/user-home?'.http_build_query(['email' => $user->getEmail()]);
+        return '/user-home?'.\http_build_query(['email' => $user->getEmail()]);
     }
 
     public function getLoginUrl($email = NULL)
     {
         $url = '/login';
         if ($email) {
-            $url .= '?'.http_build_query(['email' => $email]);
+            $url .= '?'.\http_build_query(['email' => $email]);
         }
 
         return $url;
@@ -78,7 +78,7 @@ class FixedUrlProviderStub implements UrlProvider
     {
         $url = '/register-verify-email';
         if ($email) {
-            $url .= '?'.http_build_query(['email' => $email]);
+            $url .= '?'.\http_build_query(['email' => $email]);
         }
 
         return $url;
