@@ -21,9 +21,9 @@ class UserStub extends SimpleUser
     {
         $i = new static;
         foreach ($params as $prop => $value) {
-            if ( ! property_exists($i, $prop)) {
+            if ( ! \property_exists($i, $prop)) {
                 throw new \InvalidArgumentException(
-                    'Undefined property '.$prop.' on '.get_class($i)
+                    'Undefined property '.$prop.' on '.\get_class($i)
                 );
             }
             $i->$prop = $value;
