@@ -11,8 +11,8 @@ class InsecureJSONTokenServiceStub implements EmailConfirmationTokenService
 {
     public function createToken($params)
     {
-        ksort($params);
-        return json_encode($params);
+        \ksort($params);
+        return \json_encode($params);
     }
 
     public function isValid($token, $params)
